// utils/types.ts
export type Currency = {
  id: number
  symbol: string
  name: string
  image: string
  current_price: number
  market_cap: number
  market_cap_rank: 1
  fully_diluted_valuation: number
  total_volume: number
  high_24h: number
  low_24h: number
  price_change_24h: number
  price_change_percentage_24h: number
  market_cap_change_24h: number
  market_cap_change_percentage_24h: number
  circulating_supply: number
  total_supply: number
  max_supply: number
  ath: number
  ath_change_percentage: number
  ath_date: string
  atl: number
  atl_change_percentage: number
  atl_date: string
  roi: {
    times: number
    currency: 'usd'
    percentage: number
  } | null
  last_updated: string
  price_change_percentage_24h_in_currency: number
  price_change_percentage_7d_in_currency: number
}